There are 2 ways to do the A La Russe method on Java
These are the 2 ways that allows you to do the A La Russe Multiplication method. The code in the RussianMultiplication.java file is the easier one compared to Main.java.

Both the ways requires a loop where the num1 is more than 0, and num1 /= 2 and num2 *= 2 (num1 divided by 2) (num2 multiplied by 2)
then the results of num1 which are odd is added with the corresponding value from num2.